refactor: Modularise article table component
