+++ components/ViewArticleListPage/ArticleTableHead.tsx
@@ -1,0 +1,88 @@
+import * as React from 'react'
+import Box from '@mui/material/Box'
+import TableCell from '@mui/material/TableCell'
+import TableHead from '@mui/material/TableHead'
+import TableRow from '@mui/material/TableRow'
+import TableSortLabel from '@mui/material/TableSortLabel'
+import { visuallyHidden } from '@mui/utils'
+import { Data } from './ArticleRowData'
+import { Order } from './ComparatorFunctions'
+
+interface HeadCell {
+  disablePadding: boolean
+  id: keyof Data
+  label: string
+  numeric: boolean
+}
+
+const headCells: readonly HeadCell[] = [
+  {
+    id: 'name',
+    numeric: false,
+    disablePadding: false,
+    label: 'Name',
+  },
+  {
+    id: 'date_created',
+    numeric: true,
+    disablePadding: false,
+    label: 'Date Of Creation',
+  },
+  {
+    id: 'last_modified',
+    numeric: true,
+    disablePadding: false,
+    label: 'Last Modified',
+  },
+  {
+    id: 'status',
+    numeric: true,
+    disablePadding: false,
+    label: 'Status',
+  },
+]
+
+interface ArticleTableProps {
+  onRequestSort: (
+    event: React.MouseEvent<unknown>,
+    property: keyof Data
+  ) => void
+  order: Order
+  orderBy: string
+}
+
+export const ArticleTableHead = (props: ArticleTableProps) => {
+  const { order, orderBy, onRequestSort } = props
+  const createSortHandler =
+    (property: keyof Data) => (event: React.MouseEvent<unknown>) => {
+      onRequestSort(event, property)
+    }
+
+  return (
+    <TableHead>
+      <TableRow>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            align="left"
+            padding={headCell.disablePadding ? 'none' : 'normal'}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : 'asc'}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <Box component="span" sx={visuallyHidden}>
+                  {order === 'desc' ? 'sorted descending' : 'sorted ascending'}
+                </Box>
+              ) : null}
+            </TableSortLabel>
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  )
+}
